Функциноал 99%.

+++ backend/api/v1/views.py
@@ -2,6 +2,7 @@ import os
 
 from django.contrib.auth.tokens import default_token_generator
 from django.db.models import Avg
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect
 from django_filters.rest_framework import DjangoFilterBackend
 from djoser.views import UserViewSet
@@ -16,8 +17,14 @@ from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet, ModelViewSet
 from rest_framework_simplejwt.tokens import AccessToken
 
-from backend.settings import DOMAIN_NAME
-from recipes.models import Ingredient, Recipe, ShortLink, Tag
+from backend.settings import DOMAIN_NAME, HTTPS, LOCALLY  # SHOPPING_CART
+from recipes.models import (Favoritism,
+                            Ingredient,
+                            Recipe,
+                            RecipeIngredient,
+                            ShopingCart,
+                            ShortLink,
+                            Tag)
 from recipes.utilities import get_short_link
 from users.models import Follow, User
 from .filters import IngredientSearchFilter, RecipeFilterSet
@@ -25,6 +32,7 @@ from .paginations import FootgramPageNumberPagination
 from .serializers import (AvatarSerializer,
                           IngredientSerializer,
                           RecipeReadSerializer,
+                          RecipeShortSerializer,
                           RecipeWriteSerializer,
                           TagSerializer,
                           UserFoodgramSerializer,
@@ -37,14 +45,14 @@ class UserFoodgramViewSet(UserViewSet):
 
     def get_permissions(self):
         """Переопределяет допуски к разным отдельным эндпоинтам."""
-        if self.action in ('me', 'subscriptions', 'subscribe'):
+        if self.action in ('me', 'subscriptions', 'subscribe', 'avatar'):
             self.permission_classes = [IsAuthenticated]
         return super().get_permissions()
 
     @action(detail=False, methods=['PUT', 'DELETE'], url_path='me/avatar')
     def avatar(self, request):
         """Добавляет или удаляет аватар пользователя."""
-        user = get_object_or_404(User, pk=request.user.pk)
+        user = request.user
         if request.method == 'DELETE':
             user.avatar.delete()
             user.avatar = None
@@ -59,7 +67,6 @@ class UserFoodgramViewSet(UserViewSet):
     @action(detail=False, url_path='subscriptions')
     def subscriptions(self, request):
         """Возвращает список на кого подписан."""
-        # https://www.django-rest-framework.org/api-guide/viewsets/#marking-extra-actions-for-routing тут пример есть
         following = User.objects.filter(
             following__user=request.user).order_by('id')
         page = self.paginate_queryset(following)
@@ -73,11 +80,30 @@ class UserFoodgramViewSet(UserViewSet):
         """Позволяет подписаться и отписаться на другого пользователя."""
         user = request.user
         following = get_object_or_404(User, id=id)
+        follow = Follow.objects.filter(user=user, following=following)
         if request.method == 'DELETE':
-            follow = get_object_or_404(Follow, user=user, following=following)
+            if not follow.exists():
+                return Response(
+                    {'errors': 'Вы не были подписаны на пользователя '
+                               f'{following.last_name} '
+                               f'{following.first_name}'},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
             follow.delete()
-            return Response({}, status=status.HTTP_204_NO_CONTENT)
-        Follow.objects.get_or_create(user=user, following=following)
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        if follow.exists():
+            return Response(
+                    {'errors': 'Вы уже подписаны на пользователя '
+                               f'{following.last_name} '
+                               f'{following.first_name}'},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+        if user == following:
+            return Response(
+                    {'errors': 'Нельзя подписываться на самого себя'},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+        Follow.objects.create(user=user, following=following)
         serializer = UserSubscriptionsSerializer(
             instance=following, context={'request': request}
         )
@@ -110,6 +136,15 @@ class RecipeViewSet(ModelViewSet):
     filter_backends = (DjangoFilterBackend,)
     filterset_class = RecipeFilterSet
 
+    def get_permissions(self):
+        """Переопределяет допуски к разным отдельным эндпоинтам."""
+        if self.action in (
+            'create', 'partial_update', 'destroy',
+            'favorite', 'shopping_cart', 'download_shopping_cart',
+        ):
+            self.permission_classes = [IsAuthenticated]
+        return super().get_permissions()
+
     def get_serializer_class(self):
         """В зависимости от метода запроса."""
         if self.action in ('create', 'partial_update'):
@@ -119,6 +154,80 @@ class RecipeViewSet(ModelViewSet):
     def perform_create(self, serializer):
         """Автоматически сохраняет в пользователя в поле "автор"."""
         serializer.save(author=self.request.user)
+
+    @action(detail=True, methods=['POST', 'DELETE'], url_path='favorite')
+    def favorite(self, request, pk=None):
+        """Позволяет добавлять рецепты в избранные."""
+        user = request.user
+        recipe = get_object_or_404(Recipe, pk=pk)
+        favorite = Favoritism.objects.filter(user=user, recipe=recipe)
+        if request.method == 'DELETE':
+            if not favorite.exists():
+                return Response(
+                    {'errors': f'Рецепт {recipe.name} не был в избранном.'},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+            favorite.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        if favorite.exists():
+            return Response(
+                    {'errors': f'Вы уже добавили рецепт {recipe.name} в '
+                               'избранное.'},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+        Favoritism.objects.create(user=user, recipe=recipe)
+        serializer = RecipeShortSerializer(
+            instance=recipe, context={'request': request}
+        )
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+    @action(detail=True, methods=['POST', 'DELETE'], url_path='shopping_cart')
+    def shopping_cart(self, request, pk=None):
+        """Позволяет добавлять рецепты в избранные."""
+        user = request.user
+        recipe = get_object_or_404(Recipe, pk=pk)
+        shop = ShopingCart.objects.filter(user=user, recipe=recipe)
+        if request.method == 'DELETE':
+            if not shop.exists():
+                return Response(
+                    {'errors': f'Рецепт {recipe.name} не был в корзине.'},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+            shop.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        if shop.exists():
+            return Response(
+                    {'errors': f'Вы уже добавили рецепт {recipe.name} в '
+                               'корзину.'},
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+        ShopingCart.objects.create(user=user, recipe=recipe)
+        serializer = RecipeShortSerializer(
+            instance=recipe, context={'request': request}
+        )
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+    @action(detail=False, methods=['Get'], url_path='download_shopping_cart')
+    def download_shopping_cart(self, request):
+        """Отдаст пользователю файл с его списком покупок."""
+        ingredient_list_qs = RecipeIngredient.objects.filter(
+            recipe__shop__user=request.user,
+        )
+        ingredient_list = [str(i).rsplit(' ', 1) for i in ingredient_list_qs]
+        list_for_file = {}
+        for element in ingredient_list:
+            if element[0] in list_for_file:
+                list_for_file[element[0]] += int(element[1])
+            else:
+                list_for_file[element[0]] = int(element[1])
+        file_txt = [
+            f'{ingredient.capitalize()} - {amount}\n'
+            for ingredient, amount in list_for_file.items()
+        ]
+        return HttpResponse(
+            file_txt,
+            content_type='text/plain'
+        )
 
     @action(detail=True, url_path='get-link')
     def get_link(self, request, pk=None):
@@ -136,7 +245,8 @@ class RecipeViewSet(ModelViewSet):
         else:
             short_link_recipe = short_link_recipe.first().short
         return Response(
-            {'short-link': f'https://{DOMAIN_NAME}/s/{short_link_recipe}'},
+            {'short-link': f'http{HTTPS}://{DOMAIN_NAME}'
+                           f'/s/{short_link_recipe}/'},
             status=status.HTTP_200_OK
         )
 
@@ -146,4 +256,7 @@ def redirect_short_link(request, short):
     """Обрабатывает короткие ссылки."""
     recipe_short_link = get_object_or_404(ShortLink, short=short)
     recipe_id = recipe_short_link.recipe.id
-    return redirect(f'/api/recipes/{recipe_id}/')
+    substring = '/api' if LOCALLY else ''
+    return redirect(
+        f'http{HTTPS}://{DOMAIN_NAME}{substring}/recipes/{recipe_id}/'
+    )
